Split python API system test classes and invoke ingest tasks in test
